+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Enable `check` on hotfix/ branches as well. (#36)  
+  `check`をhotfix/ブランチでも有効にする.
+
 ## [0.1.3] - 2026-03-04
 
 ### Fixed